+++ src/main/java/com/zerobase/minesweeper/controller/GameController.java
@@ -55,18 +55,4 @@ public class GameController {
     public GamerStatResponse getGamerStat(@PathVariable String gamerId) {
         return gameService.getGamerStat(gamerId);
     }
-
-    /*
-    관리자 페이지 용
-    */
-
-    //게임 삭제
-    @PreAuthorize("hasRole('ADMIN')")
-    @DeleteMapping("game/{gameId}")
-    public ResponseEntity<?> deleteGame(@PathVariable String gameId){
-        gameService.deleteGame(gameId);
-        return ResponseEntity.ok(HttpStatus.OK);
-    }
-
-
 }
