+++ src/main/java/com/zerobase/minesweeper/dto/GamersBoardInfoResponse.java
@@ -13,5 +13,6 @@ public class GamersBoardInfoResponse {
     private Long id;
     private String name;
     private String mail;
-    private LocalDate regDt;
+    private boolean isVerified;
+    private boolean isSuspend;
 }
